Move DiscoveredAPI to models
