+++ components/apimgt/org.wso2.carbon.apimgt.api/src/main/java/org/wso2/carbon/apimgt/api/model/DiscoveredAPI.java
@@ -18,9 +18,7 @@
  *
  */
 
-package org.wso2.carbon.apimgt.api;
-
-import org.wso2.carbon.apimgt.api.model.API;
+package org.wso2.carbon.apimgt.api.model;
 
 /**
  * This class represents a discovered API.
